move 'Assert' to ddal-core
