+++ ddal-ddr/pom.xml
@@ -142,6 +142,11 @@
 
     <dependencies>
         <dependency>
+            <groupId>org.hellojavaer.ddal</groupId>
+            <artifactId>ddal-core</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>org.springframework</groupId>
             <artifactId>spring-expression</artifactId>
             <version>${spring.version}</version>
