+++ ddal-ddr/src/test/java/org/hellojavaer/ddal/ddr/shard/ShardRouteContextTest.java
@@ -15,8 +15,8 @@
  */
 package org.hellojavaer.ddal.ddr.shard;
 
+import org.hellojavaer.ddal.core.utils.Assert;
 import org.hellojavaer.ddal.ddr.datasource.exception.AmbiguousDataSourceBindingException;
-import org.hellojavaer.ddal.ddr.utils.Assert;
 import org.junit.Test;
 
 /**
