+++ ddal-ddr/src/test/java/org/hellojavaer/ddal/ddr/shard/rule/SpelShardRouteRuleTest.java
@@ -15,11 +15,11 @@
  */
 package org.hellojavaer.ddal.ddr.shard.rule;
 
+import org.hellojavaer.ddal.core.utils.Assert;
 import org.hellojavaer.ddal.ddr.shard.RangeShardValue;
 import org.hellojavaer.ddal.ddr.shard.ShardRouteInfo;
 import org.hellojavaer.ddal.ddr.shard.simple.SimpleShardRouteRuleBinding;
 import org.hellojavaer.ddal.ddr.shard.simple.SimpleShardRouter;
-import org.hellojavaer.ddal.ddr.utils.Assert;
 import org.junit.Test;
 
 import java.util.*;
