+++ ddal-ddr/src/test/java/org/hellojavaer/ddal/ddr/utils/DDRJSONUtilsTest.java
@@ -15,9 +15,13 @@
  */
 package org.hellojavaer.ddal.ddr.utils;
 
+import org.hellojavaer.ddal.core.utils.Assert;
 import org.junit.Test;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
 
 /**
  *
