--- ddal-jsqlparser/src/test/java/org.hellojavaer.ddal.jsqlparser/Assert.java
@@ -1,52 +1,0 @@
-/*
- * #%L
- * ddal-jsqlparser
- * %%
- * Copyright (C) 2016 - 2017 the original author or authors.
- * %%
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation, either version 2.1 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Lesser Public License for more details.
- *
- * You should have received a copy of the GNU General Lesser Public
- * License along with this program.  If not, see
- * <http://www.gnu.org/licenses/lgpl-2.1.html>.
- * #L%
- */
-package org.hellojavaer.ddal.jsqlparser;
-
-/**
- *
- * @author <a href="mailto:hellojavaer@gmail.com">Kaiming Zou</a>,created on 16/12/2016.
- */
-public class Assert {
-
-    public static void isTrue(boolean expression) {
-        isTrue(expression, "[Assertion failed] - this expression must be true");
-    }
-
-    public static void isTrue(boolean expression, String message) {
-        if (!expression) {
-            throw new IllegalArgumentException(message);
-        }
-    }
-
-    public static void equals(String str0, String str1) {
-        if (str0 == str1) {
-            return;
-        }
-        if (str0 == null && str1 != null //
-            || str0 != null && str1 == null//
-            || !str0.equals(str1)//
-        ) {
-            throw new IllegalArgumentException("[Assertion failed] - str0 is not equal to str1. str0 is [" + str0
-                                               + "] and str1 is [" + str1 + "]");
-        }
-    }
-}
