+++ ddal-jsqlparser/src/test/java/org.hellojavaer.ddal.jsqlparser/SelectTest.java
@@ -21,9 +21,10 @@
  */
 package org.hellojavaer.ddal.jsqlparser;
 
-import org.hellojavaer.ddal.ddr.shard.ShardRouteInfo;
+import org.hellojavaer.ddal.core.utils.Assert;
 import org.hellojavaer.ddal.ddr.shard.ShardParser;
 import org.hellojavaer.ddal.ddr.shard.ShardRouteContext;
+import org.hellojavaer.ddal.ddr.shard.ShardRouteInfo;
 import org.hellojavaer.ddal.ddr.shard.simple.SimpleShardParser;
 import org.hellojavaer.ddal.ddr.sqlparse.SQLParsedResult;
 import org.junit.Before;
