--- ddal-sequence/src/main/java/org/hellojavaer/ddal/sequence/Assert.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 2017-2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.hellojavaer.ddal.sequence;
-
-/**
- *
- * @author <a href="mailto:hellojavaer@gmail.com">Kaiming Zou</a>,created on 06/01/2017.
- */
-class Assert {
-
-    public static void notNull(Object obj) {
-        isTrue(obj != null, "[Assertion failed] - this object can't be null");
-    }
-
-    public static void notNull(Object obj, String msg) {
-        isTrue(obj != null, msg);
-    }
-
-    public static void isTrue(boolean expression, String message) {
-        if (!expression) {
-            throw new IllegalArgumentException(message);
-        }
-    }
-
-}
